requirements.txt
Flask==2.2.5
gunicorn==21.2.0
requests>=2.0.0
websocket-client
pyaudio
playsound
